study-planner/backend/requirements.txt
flask
google-generativeai
python-dotenv
requests
duckduckgo-search
